Fixed digitalWrite function calls from endpoints
Within the loop function of the simpleWebServer.ino program, there is a collection of if statements that handle the turning on and off of the LEDs on the Arduino. Each if statement corresponds to a different endpoint that is callable by the created web server however, the digitalWrite statements do not accurately reflect the function of the endpoints. The endpoints have been changed to correctly reflect their intended function.